refactor(fuzzy): address PR review findings
- Move package internal/tui/fuzzy → internal/fuzzy; pure algorithm has
  no TUI dependency and should be accessible without importing tui
- Replace utf8.RuneCountInString(candidate) with len(cRunes); the rune
  slice is already computed and the utf8 import is no longer needed
- Add comment to FilterAndRank doc noting Label-only scoring is
  intentional; Sub matching deferred to a later phase
- Fix dead test row 'query not a subsequence returns zero': wantZero was
  false with wantGT -1, causing the table loop to assert nothing; set
  wantZero: true so the assertion fires as intended

Closes #69
